+++ backend/nodes/tests/test_views/test_nodes.py
@@ -1,28 +1,34 @@
 from django.urls import reverse
 
+import permissions.models
 from nodes.tests import factories
 from utils.testcases import BaseTransactionTestCase
 
 
 class NodesViewTestCase(BaseTransactionTestCase):
     def test_list(self) -> None:
+        space = factories.SpaceFactory.create(owner=self.owner_user)
         response = self.owner_client.get(reverse("nodes:nodes-list"))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 0)
-        factories.NodeFactory.create(owner=self.owner_user)
+        factories.NodeFactory.create(space=space)
         response = self.owner_client.get(reverse("nodes:nodes-list"))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 1)
 
     def test_retrieve(self) -> None:
-        node = factories.NodeFactory.create(owner=self.owner_user)
-        node_2 = factories.NodeFactory.create()
+        space = factories.SpaceFactory.create(owner=self.owner_user)
+        node = factories.NodeFactory.create(space=space)
+        node_2 = factories.NodeFactory.create(space=space)
+        node_3 = factories.NodeFactory.create(space=space)
         node.subnodes.add(node_2)
+        node_2.subnodes.add(node_3)
 
         response = self.owner_client.get(reverse("nodes:nodes-detail", args=[node.public_id]))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["id"], str(node.public_id))
         self.assertEqual(len(response.data["subnodes"]), 1)
+        self.assertTrue(response.data["subnodes"][0]["has_subnodes"])
 
         node_2.delete()
         self.assertEqual(node.subnodes.count(), 1)
@@ -39,46 +45,50 @@ class NodesViewTestCase(BaseTransactionTestCase):
         self.assertEqual(response.status_code, 405)
 
     def test_update(self) -> None:
-        node = factories.NodeFactory.create(owner=self.owner_user)
+        space = factories.SpaceFactory.create(owner=self.owner_user)
+        node = factories.NodeFactory.create(space=space)
         response = self.owner_client.patch(
             reverse("nodes:nodes-detail", args=[node.public_id]), {"name": "new name"}
         )
         self.assertEqual(response.status_code, 405)
 
     def test_delete(self) -> None:
-        node = factories.NodeFactory.create(owner=self.owner_user)
+        space = factories.SpaceFactory.create(owner=self.owner_user)
+        node = factories.NodeFactory.create(space=space)
         response = self.owner_client.delete(reverse("nodes:nodes-detail", args=[node.public_id]))
         self.assertEqual(response.status_code, 405)
 
     def test_filter_by_space(self) -> None:
         space = factories.SpaceFactory.create(owner=self.owner_user)
         another_space = factories.SpaceFactory.create(owner=self.owner_user)
-        node = factories.NodeFactory.create()
-        space.nodes.add(node)
+        node = factories.NodeFactory.create(space=space)
 
         response = self.owner_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(space.public_id)}
+            reverse("nodes:nodes-list"), {"space": str(space.public_id)}
         )
         self.assertEqual(response.status_code, 200, response.data)
         self.assertEqual(response.data["count"], 1)
         self.assertEqual(response.data["results"][0]["id"], str(node.public_id))
 
         response = self.owner_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(another_space.public_id)}
+            reverse("nodes:nodes-list"), {"space": str(another_space.public_id)}
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 0)
 
     def test_permissions(self) -> None:
+        # A node without a space should not be visible to the viewer.
         node = factories.NodeFactory.create()
 
         response = self.viewer_client.get(reverse("nodes:nodes-list"))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 0)
 
+        # Also check that the viewer can't access the node directly.
         response = self.viewer_client.get(reverse("nodes:nodes-detail", args=[node.public_id]))
         self.assertEqual(response.status_code, 403)
 
+        # Check that the viewer can't update or delete the node.
         response = self.viewer_client.patch(
             reverse("nodes:nodes-detail", args=[node.public_id]), {"name": "new name"}
         )
@@ -92,7 +102,7 @@ class NodesViewTestCase(BaseTransactionTestCase):
         space.nodes.add(node)
 
         response = self.viewer_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(space.public_id)}
+            reverse("nodes:nodes-list"), {"space": str(space.public_id)}
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 1)
@@ -108,7 +118,7 @@ class NodesViewTestCase(BaseTransactionTestCase):
         space.nodes.add(subnode)
 
         response = self.viewer_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(space.public_id)}
+            reverse("nodes:nodes-list"), {"space": str(space.public_id)}
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["count"], 2)
@@ -117,32 +127,42 @@ class NodesViewTestCase(BaseTransactionTestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.data["id"], str(subnode.public_id))
 
-        space = factories.SpaceFactory.create(owner=self.owner_user)
-        node = factories.NodeFactory.create(viewer=self.viewer_user)
-        subnode = factories.NodeFactory.create()
-        node.subnodes.add(subnode)
-        space.nodes.add(subnode, node)
-
-        # Check that we can see the nodes we have access to even though we don't have access to the
-        # space we filter by.
-        response = self.viewer_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(space.public_id)}
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data["count"], 2)
-
-        # We can also access the subnode directly because we have access to its parent node.
-        response = self.viewer_client.get(reverse("nodes:nodes-detail", args=[subnode.public_id]))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data["id"], str(subnode.public_id))
-
-        # Check that we don't see the node once it has been soft deleted.
+        # Check that we can't see the node once it has been soft-deleted
         node.delete()
         response = self.viewer_client.get(
-            reverse("nodes:nodes-list"), {"spaces": str(space.public_id)}
+            reverse("nodes:nodes-list"), {"space": str(space.public_id)}
         )
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data["count"], 0)
+        self.assertEqual(response.data["count"], 1)
+        # The list should only contain the subnode now
+        self.assertEqual(response.data["results"][0]["id"], str(subnode.public_id))
 
         response = self.viewer_client.get(reverse("nodes:nodes-detail", args=[node.public_id]))
         self.assertEqual(response.status_code, 404)
+
+    def test_requesting_node_permissions(self):
+        space = factories.SpaceFactory.create(owner=self.owner_user, viewer=self.viewer_user)
+        node = factories.NodeFactory.create(space=space)
+
+        response = self.owner_client.get(
+            reverse("nodes:nodes-detail", args=[node.public_id]), {"show_permissions": "true"}
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertSetEqual(
+            set(response.data["allowed_actions"]),
+            {
+                permissions.models.READ,
+                permissions.models.WRITE,
+                permissions.models.MANAGE,
+                permissions.models.DELETE,
+            },
+        )
+
+        response = self.viewer_client.get(
+            reverse("nodes:nodes-detail", args=[node.public_id]), {"show_permissions": "true"}
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertSetEqual(
+            set(response.data["allowed_actions"]),
+            {permissions.models.READ},
+        )