+++ backend/permissions/tests/test_views.py
@@ -6,50 +6,50 @@ from utils.testcases import BaseTransactionTestCase
 
 
 class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
-    def test_node_model_view_set(self) -> None:
-        """Test that listing the permissions is limited to the owner of the node."""
-        node = node_factories.NodeFactory.create(owner=self.owner_user)
+    def test_space_model_view_set(self) -> None:
+        """Test that listing the permissions is limited to the owner of the Space."""
+        space = node_factories.SpaceFactory.create(owner=self.owner_user)
 
         response = self.owner_client.get(
-            reverse("nodes:nodes-manage-permissions", args=[str(node.public_id)])
+            reverse("nodes:spaces-manage-permissions", args=[str(space.public_id)])
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.json()), 1)
         self.assertEqual(response.json()[0]["user"], self.owner_user.email)
 
         response = self.viewer_client.get(
-            reverse("nodes:nodes-manage-permissions", args=[str(node.public_id)])
+            reverse("nodes:spaces-manage-permissions", args=[str(space.public_id)])
         )
         self.assertEqual(response.status_code, 403)
 
         response = self.member_client.get(
-            reverse("nodes:nodes-manage-permissions", args=[str(node.public_id)])
+            reverse("nodes:spaces-manage-permissions", args=[str(space.public_id)])
         )
         self.assertEqual(response.status_code, 403)
 
         response = self.owner_client.post(
-            reverse("nodes:nodes-manage-permissions", args=[str(node.public_id)]),
+            reverse("nodes:spaces-manage-permissions", args=[str(space.public_id)]),
             data={"user": self.viewer_user.email, "role": "owner"},
         )
         self.assertEqual(response.status_code, 201)
         self.assertEqual(response.json()["user"], self.viewer_user.email)
 
         response = self.viewer_client.get(
-            reverse("nodes:nodes-manage-permissions", args=[str(node.public_id)])
+            reverse("nodes:spaces-manage-permissions", args=[str(space.public_id)])
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.json()), 2)
 
     def test_adding_and_deleting_permissions(self) -> None:
-        """Test that adding and deleting permissions is limited to the owner of the node."""
-        node = node_factories.NodeFactory.create(owner=self.owner_user)
+        """Test that adding and deleting permissions is limited to the owner of the Space."""
+        space = node_factories.SpaceFactory.create(owner=self.owner_user)
         object_permission = models.ObjectMembership.objects.get(user=self.owner_user)
 
         response = self.viewer_client.delete(
             reverse(
-                "nodes:nodes-delete-permission",
+                "nodes:spaces-delete-permission",
                 kwargs={
-                    "public_id": str(node.public_id),
+                    "public_id": str(space.public_id),
                     "permission_uuid": object_permission.public_id,
                 },
             )
@@ -58,9 +58,9 @@ class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
 
         response = self.member_client.delete(
             reverse(
-                "nodes:nodes-delete-permission",
+                "nodes:spaces-delete-permission",
                 kwargs={
-                    "public_id": str(node.public_id),
+                    "public_id": str(space.public_id),
                     "permission_uuid": str(object_permission.public_id),
                 },
             )
@@ -69,9 +69,9 @@ class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
 
         response = self.owner_client.delete(
             reverse(
-                "nodes:nodes-delete-permission",
+                "nodes:spaces-delete-permission",
                 kwargs={
-                    "public_id": str(node.public_id),
+                    "public_id": str(space.public_id),
                     "permission_uuid": str(object_permission.public_id),
                 },
             )
@@ -79,15 +79,17 @@ class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
         self.assertEqual(response.status_code, 204)
         self.assertFalse(models.ObjectMembership.objects.filter(user=self.owner_user).exists())
 
-    def test_allowed_actions_on_nodes(self) -> None:
-        """Check that we don't use N+1 queries when getting the allowed actions on a node."""
+    def test_subnodes_field_on_nodes(self) -> None:
+        """Check that we don't use N+1 queries when getting the subnodes for a node."""
 
         space = node_factories.SpaceFactory.create(owner=self.owner_user)
-        node = node_factories.NodeFactory.create(owner=self.owner_user, space=space)
-        subnodes = node_factories.NodeFactory.create_batch(10, owner=self.owner_user, space=space)
+        node = node_factories.NodeFactory.create(space=space)
+        subnodes = node_factories.NodeFactory.create_batch(10, space=space)
         node.subnodes.set(subnodes)
+        self.assertEqual(node.subnodes.count(), 10)
+        self.assertEqual(space.nodes.count(), 11)
 
-        with self.assertNumQueries(2):
+        with self.assertNumQueries(3):
             response = self.owner_client.get(
                 reverse("nodes:nodes-detail", args=[str(node.public_id)])
             )
@@ -95,10 +97,6 @@ class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
 
         response_data = response.json()
         self.assertEqual(len(response_data["subnodes"]), 10)
-        self.assertEqual(len(response_data["allowed_actions"]), 4)
-        self.assertSetEqual(
-            set(response_data["allowed_actions"]), {"read", "write", "manage", "delete"}
-        )
 
         with self.assertNumQueries(2):
             response = self.owner_client.get(reverse("nodes:nodes-list"))
@@ -106,11 +104,6 @@ class PermissionViewSetMixinTestCase(BaseTransactionTestCase):
 
         response_data = response.json()
         self.assertEqual(response_data["count"], 11)
-        for node_data in response_data["results"]:
-            self.assertEqual(len(node_data["allowed_actions"]), 4)
-            self.assertSetEqual(
-                set(node_data["allowed_actions"]), {"read", "write", "manage", "delete"}
-            )
 
     def test_allowed_actions_on_spaces(self) -> None:
         """Check that we don't use N+1 queries when getting the allowed actions on a space."""